Expose raw numeric SQLite error code to JS
In Prisma, we would like to have numeric error code as it is returned by
sqlite/libsql, rather than symbolic name. This PR adds additional
`rawCode` property to the `SqliteError` class that does just that.

This is a deviation from better-sqlite3 API: `rawCode` would be a pure
`libsql` addition. Integration tests are adjusted to check for it's
presence only on `libsql`